Move user_satisfaction into builder.
